requirements.txt
# Web scraping
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2
# Note: Selenium removed - CafeF now uses API endpoint

# Data processing
pandas==3.0.1
numpy==2.2.0
pyarrow==23.0.1

# Configuration
pyyaml==6.0.3

# Utilities
python-dateutil==2.9.0.post0
unidecode==1.4.0

# Testing
pytest==9.0.2
pytest-cov==7.0.0

# Development
ipykernel==7.2.0
jupyter==1.1.1
